+++ ChangeLog.md
@@ -18,10 +18,11 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - the new Aviron quarantine features to the :aviron module
+- the new Aviron CPU limiter features to the :aviron module
 
 ### Updated dependencies:
 
-- bumped the Aviron (ClamAV client) library to V 1.4.0
+- bumped the Aviron (ClamAV client) library to V 1.5.0